Moved edm::service::MessageLogger to plugins
Also moved pset validation class as it is only used by MessageLogger.
