+++ FWCore/MessageService/BuildFile.xml
@@ -1,7 +1,5 @@
-<use name="DataFormats/Provenance"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Utilities"/>
 <use name="tbb"/>
 <export>
