+++ FWCore/MessageService/plugins/MessageLogger.cc
@@ -15,8 +15,8 @@
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
-#include "FWCore/MessageService/interface/MessageLogger.h"
-#include "FWCore/MessageService/src/MessageServicePSetValidation.h"
+#include "FWCore/MessageService/plugins/MessageLogger.h"
+#include "FWCore/MessageService/plugins/MessageServicePSetValidation.h"
 
 #include "FWCore/MessageLogger/interface/MessageLoggerQ.h"
 #include "FWCore/MessageLogger/interface/MessageDrop.h"
